Home

MutableArithmetics.jl

MutableArithmetics.add!

MutableArithmetics.add_impl!

MutableArithmetics.add_to!

MutableArithmetics.add_to_impl!

MutableArithmetics.mul!

MutableArithmetics.mul_impl!

MutableArithmetics.mul_to!

MutableArithmetics.mul_to_impl!

MutableArithmetics.muladd!

MutableArithmetics.muladd_buf!

MutableArithmetics.muladd_buf_impl!

MutableArithmetics.muladd_buf_to!

MutableArithmetics.muladd_buf_to_impl!

MutableArithmetics.muladd_impl!

MutableArithmetics.muladd_to!

MutableArithmetics.muladd_to_impl!

MutableArithmetics.mutability

MutableArithmetics.one!

MutableArithmetics.one_impl!

MutableArithmetics.zero!

MutableArithmetics.zero_impl!

MutableArithmetics.add! — Function.

add!(a, b)

Return the sum of a and b, possibly modifying a.

source

MutableArithmetics.add_impl! — Function.

add_impl!(a, b)

Write the result of the sum of a and b to a.

source

MutableArithmetics.add_to! — Function.

add_to!(a, b, c)

Return the sum of b and c, possibly modifying a.

source

MutableArithmetics.add_to_impl! — Function.

add_to_impl!(a, b, c)

Write the result of the sum of b and c to a.

source

MutableArithmetics.mul! — Function.

mul!(a, b)

Return the product of a and b, possibly modifying a.

source

MutableArithmetics.mul_impl! — Function.

mul_impl!(a, b)

Write the result of the product of a and b to a.

source

MutableArithmetics.mul_to! — Function.

mul_to!(a, b, c)

Return the product of b and c, possibly modifying a.

source

MutableArithmetics.mul_to_impl! — Function.

mul_to_impl!(a, b, c)

Write the result of the product of b and c to a.

source

MutableArithmetics.muladd! — Function.

muladd!(a, b, c)

Return muladd(b, c, a), possibly modifying a.

source

MutableArithmetics.muladd_buf! — Function.

muladd_buf!(buf, a, b, c)

Return muladd(b, c, a), possibly modifying a and buf.

source

MutableArithmetics.muladd_buf_impl! — Function.

muladd_buf_impl!(buf, a, b, c)

Write the result of muladd(b, c, a) to a, possibly modifying buf.

source

MutableArithmetics.muladd_buf_to! — Function.

muladd_buf_to!(buf, a, b, c, d)

Return muladd(c, d, b), possibly modifying a and buf.

source

MutableArithmetics.muladd_buf_to_impl! — Function.

muladd_buf_to_impl!(buf, a, b, c, d)

Write the result of muladd(c, d, b) to a, possibly modifying buf.

source

MutableArithmetics.muladd_impl! — Function.

muladd_impl!(a, b, c)

Write the result of muladd(b, c, a) to a.

source

MutableArithmetics.muladd_to! — Function.

muladd_to!(a, b, c, d)

Return muladd(c, d, b), possibly modifying a.

source

MutableArithmetics.muladd_to_impl! — Function.

muladd_to_impl!(a, b, c, d)

Write the result of muladd(c, d, b) to a.

source

MutableArithmetics.mutability — Method.

mutability(T::Type, ::typeof(op), args::Type...)::MutableTrait

Return IsMutable to indicate that op(a::T, ::args[1], ...) returns a. That is, the result of the operation is stored in a and then a is returned. Equivalently, returns whether op_impl is supported.

source

MutableArithmetics.one! — Function.

one!(a)

Return one(a), possibly modifying a.

source

MutableArithmetics.one_impl! — Function.

one_impl!(a)

Write the result of one(a) to a.

source

MutableArithmetics.zero! — Function.

zero!(a)

Return zero(a), possibly modifying a.

source

MutableArithmetics.zero_impl! — Function.

zero_impl!(a)

Write the result of zero(a) to a.

source