MutableArithmetics.jl

MutableArithmetics._rewrite

MutableArithmetics.add!

MutableArithmetics.add_mul

MutableArithmetics.add_mul!

MutableArithmetics.add_mul_buf!

MutableArithmetics.add_mul_buf_to!

MutableArithmetics.add_mul_to!

MutableArithmetics.add_to!

MutableArithmetics.broadcast!

MutableArithmetics.broadcast_mutability

MutableArithmetics.buffered_operate!

MutableArithmetics.buffered_operate_to!

MutableArithmetics.mul!

MutableArithmetics.mul_to!

MutableArithmetics.mutability

MutableArithmetics.mutable_broadcast!

MutableArithmetics.mutable_buffered_operate!

MutableArithmetics.mutable_buffered_operate_to!

MutableArithmetics.mutable_operate!

MutableArithmetics.mutable_operate_to!

MutableArithmetics.one!

MutableArithmetics.operate!

MutableArithmetics.operate_to!

MutableArithmetics.promote_operation

MutableArithmetics.rewrite_generator

MutableArithmetics.zero!

MutableArithmetics._rewrite — Function

_rewrite(x, aff::Symbol, lcoeffs::Vector, rcoeffs::Vector, new_var::Symbol=gensym())

Return new_var, code such that code is equivalent to

new_var = aff + prod(lcoefs) * x * prod(rcoeffs)

MutableArithmetics.add! — Method

add!(a, b, ...)

Return the sum of a, b, ..., possibly modifying a.

MutableArithmetics.add_mul — Function

add_mul(a, args...)

Return a + *(args...). Note that add_mul(a, b, c) = muladd(b, c, a).

MutableArithmetics.add_mul! — Method

add_mul!(args...)

Return add_mul(args...), possibly modifying args[1].

MutableArithmetics.add_mul_buf! — Method

add_mul_buf!(buffer, args...)

Return add_mul(args...), possibly modifying args[1] and buffer.

MutableArithmetics.add_mul_buf_to! — Method

add_mul_buf_to!(buffer, output, args...)

Return add_mul(args...), possibly modifying output and buffer.

MutableArithmetics.add_mul_to! — Method

add_mul_to!(output, args...)

Return add_mul(args...), possibly modifying output.

MutableArithmetics.add_to! — Method

add_to!(a, b, c)

Return the sum of b and c, possibly modifying a.

MutableArithmetics.broadcast! — Method

broadcast!(op::Function, args...)

Returns the value of broadcast(op, args...), possibly modifying args[1].

MutableArithmetics.broadcast_mutability — Method

broadcast_mutability(T::Type, ::typeof(op), args::Type...)::MutableTrait

Return IsMutable to indicate an object of type T can be modified to be equal to broadcast(op, args...).

MutableArithmetics.buffered_operate! — Method

buffered_operate!(buffer, op::Function, args...)

Returns the value of op(args...), possibly modifying buffer.

MutableArithmetics.buffered_operate_to! — Method

buffered_operate_to!(buffer, output, op::Function, args...)

Returns the value of op(args...), possibly modifying buffer and output.

MutableArithmetics.mul! — Method

mul!(a, b, ...)

Return the product of a, b, ..., possibly modifying a.

MutableArithmetics.mul_to! — Method

mul_to!(a, b, c)

Return the product of b and c, possibly modifying a.

MutableArithmetics.mutability — Method

mutability(T::Type, ::typeof(op), args::Type...)::MutableTrait

Return IsMutable to indicate an object of type T can be modified to be equal to op(args...).

MutableArithmetics.mutable_broadcast! — Function

mutable_broadcast!(op::Function, args...)

Modify the value of args[1] to be equal to the value of broadcast(op, args...). Can only be called if mutability(args[1], op, args...) returns true.

MutableArithmetics.mutable_buffered_operate! — Method

mutable_buffered_operate!(buffer, op::Function, args...)

Modify the value of args[1] to be equal to the value of op(args...), possibly modifying buffer. Can only be called if mutability(args[1], op, args...) returns true.

MutableArithmetics.mutable_buffered_operate_to! — Method

mutable_buffered_operate_to!(buffer, output, op::Function, args...)

Modify the value of output to be equal to the value of op(args...), possibly modifying buffer. Can only be called if mutability(output, op, args...) returns true.

MutableArithmetics.mutable_operate! — Method

mutable_operate!(op::Function, args...)

Modify the value of args[1] to be equal to the value of op(args...). Can only be called if mutability(args[1], op, args...) returns true.

MutableArithmetics.mutable_operate_to! — Method

mutable_operate_to!(output, op::Function, args...)

Modify the value of output to be equal to the value of op(args...). Can only be called if mutability(output, op, args...) returns true.

MutableArithmetics.one! — Method

one!(a)

Return one(a), possibly modifying a.

MutableArithmetics.operate! — Method

operate!(op::Function, args...)

Returns the value of op(args...), possibly modifying args[1].

MutableArithmetics.operate_to! — Method

operate_to!(output, op::Function, args...)

Returns the value of op(args...), possibly modifying output.

MutableArithmetics.promote_operation — Function

promote_operation(op::Function, ArgsTypes::Type...)

Returns the type returned to the call operate(op, args...) where the types of the arguments args are ArgsTypes.

MutableArithmetics.rewrite_generator — Method

rewrite_generator(expr::Expr, inner::Function)

Rewrites the generator statements expr and returns a properly nested for loop with nested filters as specified.

Examples

julia> using MutableArithmetics

julia> MutableArithmetics.rewrite_generator(:(i for i in 1:2 if isodd(i)), i -> :($i + 1))
:(for $(Expr(:escape, :(i = 1:2)))
      if $(Expr(:escape, :(isodd(i))))
          i + 1
      end
  end)

MutableArithmetics.zero! — Method

zero!(a)

Return zero(a), possibly modifying a.